gps functionality done
so, this is pretty much the second commit to the gps branch (deleted),
but i didnt want to commit to that branch and then merge it to master
due to some android gradle problems. anyway, this commit is the
equivalent of making a new branch and merging it. GPS works!!!!!! still
need to test if GPS works with android versions that actually let you do
the whole request permission thing